% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bRexit.R
\name{run.bRexit}
\alias{run.bRexit}
\title{bRexit}
\usage{
run.bRexit()
}
\description{
This function runs an infinite loop. If you run it, it will slowly but surely eat up all your memory causing your system to halt. It is entirely pointless.
}
